# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Pet Helper API Documentation
  version: 1.0.0
  description: Tài liệu API cho dự án Pet Helper. Hỗ trợ xác thực bằng JWT.
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/products:
    get:
      summary: Get all products
      tags: [Products]
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        200: { description: List of products }
    post:
      summary: Add product
      tags: [Products]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: Product added }
  /api/v1/products/{id}:
    get:
      summary: Get product details
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Product details }
  /api/v1/products/{id}/like-status:
    get:
      summary: Check like status
      tags: [Products]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: user_id
          schema: { type: string }
      responses:
        200: { description: Like status }
  /api/v1/products/{id}/like:
    post:
      summary: Toggle like for a product
      tags: [Products]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Like toggled }
  /api/v1/products/admin/all:
    get:
      summary: Get all products (Admin)
      tags: [Products]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: List of all products }

  # --- Reviews ---
...